Needs another level of type caching, to ensure that there "is" only one of each type around.  The _cffi_backend code depends on that.

+++ new/parse_c_type.h
@@ -71,7 +71,7 @@ struct _cffi_struct_union_s {
     int type_index;          // -> _cffi_types, on a OP_STRUCT_UNION
     int flags;               // CT_UNION?  CT_IS_OPAQUE?
     size_t size;
-    size_t alignment;
+    int alignment;
     int first_field_index;   // -> _cffi_fields array
     int num_fields;
 };
